refactor: Entry API cleanup
Not only does the parse_entry() machinery belong to Entry(), it is also
a little cleaner in that it removes thee paths for setting defaults from
the processing.

Also, it never made sense to instantiate objects with their text *and*
their repeated attributes.  The assumption that parse_entry() would be
the only way create objects existed from creation, and the code should
have matched that.  This removes that misstep.