perf: Use single re.split() to break up input
A different approach this time.  Instead of repeated searches across the
input text, we could create a word list to poke at instead.

I suspect this will be considerably faster than both the original and
the index+re.findall() from the previous commit.  However, it will need
to be noticeably Real World faster on my archive to make it worth the
cost in readability.