refactor: Lazily extract Entry properties
This makes object creation practically instantaneous, but it also makes
it a little harder to reason about performance when operating on
Entries().

It does present an enormous speed up in basic views for tools I’ve been
playing with, but it would need a significant shake down before I’d be
happy merging it.

Refs #3.