fix: Block modification of computed properties
Attempts to set computed properties will likely leave the object out of
sync, and should be blocked.

It is questionable whether some properties, namely completion and
priority, could be set manually.  However, the other properties raise
only questions about style given that we’re treating the Entry as
free-form text.

No current consumers rely on this behaviour, so this is not a breaking
change.