chore: Use manual IntEnum setup for typechecker support
This *awful* block exists because if we use the dynamic interface for
IntEnum setup then we can't use the typechecker to guarantee our usage
at call sites.

The one good thing about this method is that it reminds us how useful
vim's "g C-x" can be when combined with nrformats=alpha.