fix: Correct handling of tasks without completion date
It probably says an awful lot about me — none of which is good — but it
has taken a year for me to catch this bug.  In over a year of heavy use,
I've never had a task marked as complete without a completion date.

[And looking at my data there is not a single entry that didn’t have a
``started`` key added either.  Fastidious is hardly ever a compliment.]